#13 HL7 - Corrected exception handling.
 * From now on success status (CA) will be send back in ACK in case of PatientNotFound and StayNotFound exceptions.

When HL7 gateway starts work, it may receive patient discharge (ADT^08) or details update (ADT^03) message before receive patient registration (ADT^01) message. In that case these messages can not be processed correctly,  so PatientNotFoundException or  StayNotFoundException are thrown, but still response (ACK) with success status will be send back (CA). The reason for that is clients HL7 gateway, which does not expect error in that situation, so to achieve seamless integration between gateways these exceptions are merely logged.